[TS migration] Migrate 'withViewportOffsetTop.js' HOC to TypeScript
